AB6. As a member of an organization, I want to delete other members, so that they no longer have access to the organization resources.

+++ app/imports/api/departments/methods.js
@@ -4,7 +4,7 @@ import { SimpleSchema } from 'meteor/aldeed:simple-schema';
 import { DepartmentSchema } from './department-schema';
 import DepartmentService from './department-service.js';
 import { IdSchema } from '../schemas.js';
-import { checkUserId } from '../checkers.js';
+
 
 export const insert = new ValidatedMethod({
   name: 'Departments.insert',
@@ -12,9 +12,11 @@ export const insert = new ValidatedMethod({
   validate: DepartmentSchema.validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot create a department'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot create a department'
+      );
+    }
 
     return DepartmentService.insert(doc);
   }
@@ -28,9 +30,11 @@ export const update = new ValidatedMethod({
   }]).validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot update a department'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot update a department'
+      );
+    }
 
     return DepartmentService.update(doc);
   }
@@ -42,9 +46,11 @@ export const remove = new ValidatedMethod({
   validate: IdSchema.validator(),
 
   run(doc) {
-    checkUserId(
-      this.userId, 'Unauthorized user cannot remove a department'
-    );
+    if (!this.userId) {
+      throw new Meteor.Error(
+        403, 'Unauthorized user cannot remove a department'
+      );
+    }
 
     return DepartmentService.remove(doc);
   }
